10/12 Revision
A re-recorded my intro and body text now that I do not have a cold.  I made sure to annunciate better and talk loudly, knowing that I can envelope my speaking to turn down that volume if necessary.  My plan for class today is to work on the tuba music, placing it in appropriate spots throughout the text.  I also want to play around with adding sounds of a kitchen to the background.  I may need to do some noise cancelling on my talking piece, and ask about how to approach a few instances of reverberations (if thats what you call them) during my speaking.